+++ src/Exception/ConfigException.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SugiPHP\Config\Exception;
+
+use RuntimeException;
+
+/**
+ * Base exception class.
+ * Specific exceptions are based on this exception/
+ */
+class ConfigException extends RuntimeException
+{
+}