+++ src/Exception/FileException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SugiPHP\Config\Exception;
+
+/**
+ * Exception for signaling file errors.
+ */
+class FileException extends ConfigException
+{
+}