+++ src/Exception/ParserException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SugiPHP\Config\Exception;
+
+/**
+ * Exception for signaling parser errors.
+ */
+class ParserException extends ConfigException
+{
+}